fluenttest/fluenttest-framework/requirements.txt
# FluentTest Framework Dependencies
# Core Appium dependencies
Appium-Python-Client==3.1.0
selenium==4.15.0

# Data processing and utilities
numpy==1.24.3
pandas==2.0.3

# Testing framework
pytest==7.4.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0

# Optional: For enhanced functionality
requests==2.31.0
pillow==10.0.0
